Add files via upload
This scene demonstrates the conveyor belt capability. In the simulation, a detectable block is placed on a conveyor belt. The conveyor belt the parent of a ray proximity sensor. Once the simulation starts, the block moves along the belt until it hits the proximity sensor. Once sensed, the belt velocity is changed to zero. 

A dummy can be placed on the block, to be used as a "waypoint" for the end effector to be sent to. This would likely be best for a demonstration of inverse kinematics.